Added Patient, Procedure and join monster
Schema definitions in .graphql are no longer used because we need to
define schema in the code in order to decorate the types with
JoinMonster properties.

JoinMonster takes care of computing JOIN clauses for our SQL. The
tradeoff is that we sacrifice ability to build a query as an object and
need to depend on string concatenation and manual sanitization of
inputs.